extracted sort function and options into utils file
